requirements.txt
flask
pystray
pillow
waitress